Extract module `settings`
Cleans up the code a bit and separates concerns.
